big update

+++ app/src/main/java/com/example/giua_ki/model/SizeModel.java
@@ -20,8 +20,8 @@ public class SizeModel {
         this.size = size;
     }
 
-    public double getPrice() {
-        return price;
+    public int getPrice() {
+        return (int) price;
     }
 
     public void setPrice(double price) {
